+++ packages/stream-model-instance-handler/src/model-instance-document-handler.ts
@@ -22,6 +22,7 @@ import { StreamID } from '@ceramicnetwork/streamid'
 import { SchemaValidation } from './schema-utils.js'
 import { Model } from '@ceramicnetwork/stream-model'
 import { applyAnchorCommit } from '@ceramicnetwork/stream-handler-common'
+import { toString } from 'uint8arrays'
 
 // Hardcoding the model streamtype id to avoid introducing a dependency on the stream-model package
 const MODEL_STREAM_TYPE_ID = 2
@@ -94,11 +95,11 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
    */
   async _applyGenesis(commitData: CommitData, context: StreamReaderWriter): Promise<StreamState> {
     const payload = commitData.commit
-    const { controllers, model, context: ctx } = payload.header
+    const { controllers, model, context: ctx, unique } = payload.header
     const controller = controllers[0]
     const modelStreamID = StreamID.fromBytes(model)
     const streamId = new StreamID(ModelInstanceDocument.STREAM_TYPE_ID, commitData.cid)
-    const metadata = { controllers: [controller], model: modelStreamID } as StreamMetadata
+    const metadata = { controllers: [controller], model: modelStreamID, unique } as StreamMetadata
     if (ctx) {
       metadata.context = StreamID.fromBytes(ctx)
     }
@@ -124,7 +125,7 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
         modelStreamID,
         streamId
       )
-    } else if (payload.data || payload.header.unique) {
+    } else if (payload.data) {
       throw Error('ModelInstanceDocument genesis commit with content must be signed')
     }
 
@@ -135,7 +136,7 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
 
     return {
       type: ModelInstanceDocument.STREAM_TYPE_ID,
-      content: payload.data || {},
+      content: payload.data || null,
       metadata,
       signature: SignatureStatus.SIGNED,
       anchorStatus: AnchorStatus.NOT_REQUESTED,
@@ -180,10 +181,19 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
       )
     }
 
-    const oldContent = state.content
+    const oldContent = state.content ?? {}
     const newContent = jsonpatch.applyPatch(oldContent, payload.data).newDocument
     const modelStream = await context.loadStream<Model>(metadata.model)
+<<<<<<< HEAD
     await this._validateContent(context, modelStream, newContent, false, payload)
+=======
+    await this._validateContent(context, modelStream, newContent, false)
+    await this._validateUnique(
+      modelStream,
+      metadata as unknown as ModelInstanceDocumentMetadata,
+      newContent
+    )
+>>>>>>> develop
 
     state.signature = SignatureStatus.SIGNED
     state.anchorStatus = AnchorStatus.NOT_REQUESTED
@@ -231,7 +241,11 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
     genesis: boolean,
     payload?: Payload
   ): Promise<void> {
-    if (genesis && model.content.accountRelation.type === 'single') {
+    if (
+      genesis &&
+      (model.content.accountRelation.type === 'single' ||
+        model.content.accountRelation.type === 'set')
+    ) {
       if (content) {
         throw new Error(
           `Deterministic genesis commits for ModelInstanceDocuments must not have content`
@@ -347,6 +361,41 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
       if (mutated) {
         throw new Error(`Immutable field "${lockedField}" cannot be updated`)
       }
+    }
+  }
+
+  /*
+   * Validates the ModelInstanceDocument unique constraints against the Model definition.
+   * @param model - model that this ModelInstanceDocument belongs to
+   * @param metadata - ModelInstanceDocument metadata to validate
+   * @param content - ModelInstanceDocument content to validate
+   */
+  async _validateUnique(
+    model: Model,
+    metadata: ModelInstanceDocumentMetadata,
+    content: Record<string, unknown> | null
+  ): Promise<void> {
+    // Unique field validation only applies to the SET account relation
+    if (model.content.accountRelation.type !== 'set') {
+      return
+    }
+    if (metadata.unique == null) {
+      throw new Error('Missing unique metadata value')
+    }
+    if (content == null) {
+      throw new Error('Missing content')
+    }
+
+    const unique = model.content.accountRelation.fields
+      .map((field) => {
+        const value = content[field]
+        return value ? value.toString() : ''
+      })
+      .join('|')
+    if (unique !== toString(metadata.unique)) {
+      throw new Error(
+        'Unique content fields value does not match metadata. If you are trying to change the value of these fields, this is causing this error: these fields values are not mutable.'
+      )
     }
   }
 }