+++ packages/stream-tests/src/__tests__/deterministic-model-instance-document.test.ts
@@ -16,7 +16,7 @@ import { Model, ModelDefinition } from '@ceramicnetwork/stream-model'
 const CONTENT0 = { myData: 0 }
 const CONTENT1 = { myData: 1 }
 
-const MODEL_DEFINITION: ModelDefinition = {
+const MODEL_DEFINITION_SINGLE: ModelDefinition = {
   name: 'MyModel',
   version: '1.0',
   accountRelation: { type: 'single' },
@@ -32,6 +32,30 @@ const MODEL_DEFINITION: ModelDefinition = {
       },
     },
     required: ['myData'],
+  },
+}
+
+const MODEL_DEFINITION_SET: ModelDefinition = {
+  name: 'MyModel',
+  version: '2.0',
+  interface: false,
+  implements: [],
+  accountRelation: { type: 'set', fields: ['unique'] },
+  schema: {
+    $schema: 'https://json-schema.org/draft/2020-12/schema',
+    type: 'object',
+    additionalProperties: false,
+    properties: {
+      unique: {
+        type: 'string',
+        minLength: 3,
+      },
+      other: {
+        type: 'string',
+        minLength: 3,
+      },
+    },
+    required: ['unique'],
   },
 }
 
@@ -65,7 +89,7 @@ describe('ModelInstanceDocument API http-client tests', () => {
     ceramic = new CeramicClient(apiUrl)
     ceramic.did = core.did
 
-    model = await Model.create(ceramic, MODEL_DEFINITION)
+    model = await Model.create(ceramic, MODEL_DEFINITION_SINGLE)
     midMetadata = { model: model.id }
   }, 12000)
 
@@ -131,6 +155,45 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(doc2.id.toString()).toEqual(doc1.id.toString())
     expect(doc2.content).toEqual(doc1.content)
   })
+
+  describe('Using custom unique metadata values', () => {
+    test('Cannot replace unique content fields values', async () => {
+      const model = await Model.create(ceramic, MODEL_DEFINITION_SET)
+
+      const doc = await ModelInstanceDocument.set(ceramic, { model: model.id }, ['foo'])
+
+      await expect(() => doc.replace({ unique: 'test' })).rejects.toThrow(
+        /Unique content fields value does not match metadata/
+      )
+
+      await doc.replace({ unique: 'foo', other: 'bar' })
+      expect(doc.content).toEqual({ unique: 'foo', other: 'bar' })
+    })
+
+    test('Cannot set content that does not pass validation', async () => {
+      const model = await Model.create(ceramic, MODEL_DEFINITION_SET)
+
+      const doc = await ModelInstanceDocument.set(ceramic, { model: model.id }, ['a'])
+      await expect(() => doc.replace({ unique: 'a', other: 'test' })).rejects.toThrow(
+        /Validation Error: data\/unique must NOT have fewer than 3 characters/
+      )
+    })
+
+    test('Can create, load and update unique documents', async () => {
+      const model = await Model.create(ceramic, MODEL_DEFINITION_SET)
+
+      const doc1 = await ModelInstanceDocument.set(ceramic, { model: model.id }, ['foo'])
+      await doc1.replace({ unique: 'foo', other: 'test' })
+
+      const doc2 = await ModelInstanceDocument.set(ceramic, { model: model.id }, ['foo'])
+      expect(doc2.id.toString()).toBe(doc1.id.toString())
+      expect(doc2.content).toEqual({ unique: 'foo', other: 'test' })
+
+      const doc3 = await ModelInstanceDocument.set(ceramic, { model: model.id }, ['bar'])
+      expect(doc3.id.toString()).not.toBe(doc1.id.toString())
+      expect(doc3.content).toBeNull()
+    })
+  })
 })
 
 describe('ModelInstanceDocument API multi-node tests', () => {
@@ -153,7 +216,7 @@ describe('ModelInstanceDocument API multi-node tests', () => {
     ceramic0 = await createCeramic(ipfs0)
     ceramic1 = await createCeramic(ipfs1)
 
-    model = await Model.create(ceramic0, MODEL_DEFINITION)
+    model = await Model.create(ceramic0, MODEL_DEFINITION_SINGLE)
     midMetadata = { model: model.id }
   }, 12000)